feat(v3): port AEROVAULT3 header, manifest and block assembly
Add the on-disk format of the AEROVAULT3 container (T3 of the crate rev. 4
migration), ported byte-for-byte from the AeroFTP app:

- v3::format: the fixed 1024-byte VaultHeaderV3 (exact field offsets), to_bytes
  / from_bytes with the reserved-region and magic/version/length checks, and
  HMAC-SHA512 integrity over the full header with the MAC field zeroed; plus
  derive_keks (base KEK -> master/MAC KEK via HKDF-SHA256).
- v3::manifest: the AlgorithmSpec / WrapperManifest / ManifestEntryV3 /
  ChunkRecordV3 / VaultManifestV3 / ExtensionEntryV3 serde structs (field order
  + skip-if-none preserved), default_wrappers, empty_manifest, CDC-bounds and
  zstd-level resolution, manifest AEAD encrypt/decrypt, block AAD, next block
  index, now_iso.
- v3::block: the [block_len u64 LE][ciphertext] data section, a bounded section
  reader, and build_file_bytes (header | data | manifest | extension dir | ext
  payloads) with offsets recomputed and the header MAC stamped. The empty
  extension directory is the 2-byte JSON "[]" (byte-critical). EC-metadata
  auto-injection is deferred to T6 (rev. 3 carries no extensions).

Frozen header-layout digest, manifest round-trip, MAC tamper-rejection and
assembly-layout tests; clippy -D warnings clean. Full app cross-open: T5.

+++ aerovault/src/v3/packing.rs
@@ -118,7 +118,7 @@ mod tests {
     fn pack_flushes_at_target() {
         // Three ~half-target small files -> flush after the first crosses target
         // would need >= PACK_TARGET; build members each just under threshold.
-        let chunk = (PACK_SMALL_FILE_THRESHOLD - 1) as usize; // < threshold => small
+        let chunk = PACK_SMALL_FILE_THRESHOLD - 1; // < threshold => small
         let small = vec![
             ("a".to_string(), vec![0u8; chunk]),
             ("b".to_string(), vec![0u8; chunk]),
